+++ opentelemetry-api/src/opentelemetry/baggage/propagation/__init__.py
@@ -21,9 +21,8 @@ from opentelemetry.context.context import Context
 from opentelemetry.trace.propagation import textmap
 
 
-class BaggagePropagator(textmap.TextMapPropagator):
-    """Extracts and injects Baggage which is used to annotate telemetry.
-    """
+class W3CBaggagePropagator(textmap.TextMapPropagator):
+    """Extracts and injects Baggage which is used to annotate telemetry."""
 
     MAX_HEADER_LENGTH = 8192
     MAX_PAIR_LENGTH = 4096