+++ opentelemetry-api/src/opentelemetry/propagate/__init__.py
@@ -25,7 +25,7 @@ would instantiate
 ``opentelemetry.propagators.composite.CompositeHTTPPropagator`` with 2
 propagators, one of type
 ``opentelemetry.trace.propagation.tracecontext.TraceContextTextMapPropagator``
-and other of type ``opentelemetry.baggage.propagation.BaggagePropagator``.
+and other of type ``opentelemetry.baggage.propagation.W3CBaggagePropagator``.
 Notice that these propagator classes are defined as
 ``opentelemetry_propagator`` entry points in the ``setup.cfg`` file of
 ``opentelemetry``.
@@ -128,7 +128,8 @@ try:
 
     # Single use variable here to hack black and make lint pass
     environ_propagators = environ.get(
-        OTEL_PROPAGATORS, "tracecontext,baggage",
+        OTEL_PROPAGATORS,
+        "tracecontext,baggage",
     )
 
     for propagator in environ_propagators.split(","):
@@ -149,6 +150,8 @@ def get_global_textmap() -> textmap.TextMapPropagator:
     return _HTTP_TEXT_FORMAT
 
 
-def set_global_textmap(http_text_format: textmap.TextMapPropagator,) -> None:
+def set_global_textmap(
+    http_text_format: textmap.TextMapPropagator,
+) -> None:
     global _HTTP_TEXT_FORMAT  # pylint:disable=global-statement
     _HTTP_TEXT_FORMAT = http_text_format  # type: ignore